:wrench: Make icon picker implementation icon-theme dependent
