+++ src/renderer/editor/properties/style/index.scss
@@ -8,7 +8,6 @@
 // SPDX-FileCopyrightText: Simon Schneegans <code@simonschneegans.de>
 // SPDX-License-Identifier: MIT
 
-@import './icon-picker.scss';
 @import './text-picker.scss';
 
 #kando-menu-properties-area {
