+++ src/renderer/editor/toolbar/trash-tab.ts
@@ -15,7 +15,7 @@ import { ToolbarDraggable } from './toolbar-draggable';
 import { IMenu, IMenuItem, deepCopyMenu, deepCopyMenuItem } from '../../../common';
 import { IEditorMenuItem } from '../common/editor-menu-item';
 import { ItemTypeRegistry } from '../../../common/item-type-registry';
-import { IconThemeRegistry } from '../../../common/icon-theme-registry';
+import { IconThemeRegistry } from '../../icon-themes/icon-theme-registry';
 import { IDraggable } from '../common/draggable';
 import { DnDManager } from '../common/dnd-manager';
 
@@ -97,9 +97,10 @@ export class TrashTab extends DropTargetTab {
           name: menu.root.name,
           description:
             (this.showShortcutIDs ? menu.shortcutID : menu.shortcut) || 'Not bound.',
-          icon: IconThemeRegistry.getInstance()
-            .getTheme(menu.root.iconTheme)
-            .createDiv(menu.root.icon).outerHTML,
+          icon: IconThemeRegistry.getInstance().createIcon(
+            menu.root.iconTheme,
+            menu.root.icon
+          ).outerHTML,
           index,
         };
       }
@@ -111,9 +112,8 @@ export class TrashTab extends DropTargetTab {
         isMenu: false,
         name: item.name,
         description: typeInfo?.getDescription(item),
-        icon: IconThemeRegistry.getInstance()
-          .getTheme(item.iconTheme)
-          .createDiv(item.icon).outerHTML,
+        icon: IconThemeRegistry.getInstance().createIcon(item.iconTheme, item.icon)
+          .outerHTML,
         index,
       };
     });
