+++ src/renderer/icon-themes/emoji-theme.ts
@@ -11,13 +11,13 @@
 import { matchSorter } from 'match-sorter';
 import emojis from 'emojilib';
 
-import { IIconTheme } from '../icon-theme-registry';
+import { IconListTheme } from './icon-list-theme';
 
 /**
  * This class implements an icon theme that uses emojis as icons. It uses the `emojilib`
  * package to get a list of emojis and their descriptions.
  */
-export class EmojiTheme implements IIconTheme {
+export class EmojiTheme extends IconListTheme {
   /**
    * This array contains all emojis and their descriptions. Each inner array contains the
    * emoji itself as first element and all descriptions as following elements.
@@ -25,6 +25,8 @@ export class EmojiTheme implements IIconTheme {
   private icons: Array<Array<string>> = [];
 
   constructor() {
+    super();
+
     // Transform the emoji lib object into a nested array where each inner array contains
     // the emoji itself and all descriptions.
     this.icons = Object.entries(emojis).map(([emoji, descriptors]) => {
@@ -56,7 +58,7 @@ export class EmojiTheme implements IIconTheme {
    * @param icon One of the icons returned by `listIcons`.
    * @returns A div element that contains the icon.
    */
-  createDiv(icon: string) {
+  createIcon(icon: string) {
     const containerDiv = document.createElement('div');
     containerDiv.classList.add('icon-container');
